api/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
coverage==7.5.2
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
python-dotenv==1.0.1
Werkzeug==3.0.3
